Scaling prices of independent trades (#20)
The new price check in the driver are sometimes preventing cow dex solver from submitting solutions:
If there are independent trades, e.g. DAI -> USDC and ETH -> GNO, the prices between DAI and GNO still need to satisfy the price check in the driver. Hence, the prices of unrelated trades still need to consider the external prices provided for the auction for the setting of the new price with the correct ratio. This was not done before.

I also reactivated all the old test, in order to have more control over the insert_new_price function.

Testplan:
unit tests only.

Note:
There are several refactors overdue, e.g. the separation of the logic of inserting_new_price and the merge of splitted_trade_amounts and the SwapResponse traded amounts. But this will happen in a future PR.