Add an absolute slippage cap (#35)
Gnosis' CowDexSolver is causing some severe negative slippage from time to time (especially on large trades) as we are only limiting slippage with a percentage and not an absolute amount. This PR adds the latter.

The logic here is heavily borrowed from https://github.com/cowprotocol/services/pull/591 (stripped off a few things like solver dependent value etc that are not needed here). We introduce a SlippageCalculator component that can compute  slippage based on absolute and relative bounds (all we need is the external prices which we already have).

## Test Plan

Unit tests from the code (I copy pasted the computation logic)

## Release Plan

relative slippage stays at 0.1% default, other values have to be configured in our deployment. I will take care of this (first in staging and then prod).